[irods/irods 8653] imiscsvrinfo: Do not segfault when connected to 4.3.3 and earlier
Before this commit, imiscsvrinfo would segfault when attempting to
deserialize the response from a server older than 4.3.4. This was due
to 4.3.4 servers and later expecting the byte stream to include data for
certificate information.

This problem is fixed by overriding the packing instruction used for
deserialization of the response. This only applies when connected to a
server that's older than 4.3.4.